shell sort

+++ sort.h
@@ -25,7 +25,7 @@ void _swap(int *a, int *b);
 void insertion_sort_list(listint_t **list);
 void selection_sort(int *array, int size);
 void quick_sort(int *array, int size);
-void shell_sort(int *array, int size);
+void shell_sort(int *array, size_t size);
 void cocktail_sort_list(listint_t **list);
 void counting_sort(int *array, int size);
 void merge_sort(int *array, int size);
